dataset split

+++ test/get_train_test.py
@@ -1,7 +1,6 @@
 import torch
 from torch.utils.data import Dataset
 from torch.utils.data.sampler import BatchSampler
-from config import *
 import shutil
 
 import os
@@ -10,6 +9,10 @@ import numpy as np
 import random
 
 from sklearn.utils import shuffle
+
+from config import *
+from feature_extraction import *
+
 
 OBJ_DICT = {
     'top_razor-core-x': '2-1',
@@ -23,7 +26,7 @@ OBJ_DICT = {
     'dehumidifier-top': '4-1',
     'dehumidifier-body': '4-2',
 
-    'top2_dell-inspire-case-right-side': '  ',
+    'top2_dell-inspire-case-right-side': '5-1',
     'top1_dell-inspire-case-left-side': '5-2',
     'side_dell-inspire-case-right-side': '5-3',
     'dell-inspire-case-front': '5-4',
@@ -103,29 +106,32 @@ LABEL_DICT = {
 }
 
 
-def get_target_file_name(sound_file_name, i):
-    target_file_name = OBJ_DICT[sound_file_name] + str(i) + '-1-1.npy'
+def get_target_file_name(obj, i):
+    target_file_name = obj.exp_name + '-' + str(obj.start_num + i) + '-1-1.npy'
     return target_file_name
 
 
-# 将声音文件从源文件夹复制到目标文件夹，并改名
-def copy_file(obj_list):
+# 将仿真声音文件从源文件夹复制到目标文件夹，并改名
+def copy_sim_file(obj_list, overwrite=0):
     root_path = os.path.join(global_var.DATASET, global_var.SOUND_ALL)
     target_path = os.path.join(global_var.DATASET, global_var.RAW_DATA, 'sim_data')
+    if not os.path.exists(target_path):
+        os.makedirs(target_path)
     for obj in obj_list.obj_list:
         obj_path = os.path.join(root_path, '%s/mat-%s' % (obj.obj_file_name, obj.material_file_name))
         sound_path = os.path.join(obj_path, obj.sound_prefix + obj.obj_file_name + '_sound' + obj.sound_postfix)
         total_num = np.prod(obj.num)
         for i in range(total_num):
-            sound_file_name = obj.sound_prefix + obj.obj_file_name + '-' + str(obj.start_num + i) + '.npy'
+            sound_file_name = obj.sound_prefix + obj.obj_file_name + '-' + str(i + 1) + '.npy'
             sound_file_path = os.path.join(sound_path, sound_file_name)
-            target_file_name = get_target_file_name(obj.sound_prefix + obj.obj_file_name, obj.start_num + i)
+            target_file_name = get_target_file_name(obj, i)
             target_file_path = os.path.join(target_path, target_file_name)
-            shutil.copyfile(sound_file_path, target_file_path)
+            if not os.path.exists(target_file_path) or overwrite:
+                shutil.copyfile(sound_file_path, target_file_path)
 
 
 def get_label(file_name):
-    res = LABEL_DICT.get('-'.join(file_name.split('-')[0:2]), "Not Found!")
+    res = LABEL_DICT.get('-'.join(file_name.split('-')[:2]), "Not Found!")
     if res == "Not Found!":
         print(file_name)
         return np.int64(0)
@@ -133,49 +139,134 @@ def get_label(file_name):
         return res
 
 
-# 将所有声音npy数据读取到内存
-def get_dataset(obj_list):
-    # 读取仿真数据
-    sim_data = []
-    sim_label = []
-    sim_path = os.path.join(global_var.DATASET, global_var.RAW_DATA, 'sim_data')
-    for obj in obj_list.obj_list:
-        total_num = np.prod(obj.num)
-        for i in range(total_num):
-            sim_file_name = get_target_file_name(obj.sound_prefix + obj.obj_file_name, obj.start_num + i)
-            sim_file_path = os.path.join(sim_path, sim_file_name)
-            sim_file_data = np.load(sim_file_path)
-            sim_data.append(sim_file_data)
-            label = get_label(sim_file_name)
-            sim_label.append(label)
+# 处理raw_data中的数据
+def raw_to_feature_dataset(obj_list, target_path, transform, overwrite=0):
+    for path in ['sim_data', 'target_data']:
+        raw_path = os.path.join(global_var.DATASET, global_var.RAW_DATA, path)
+        feature_path = os.path.join(global_var.DATASET, global_var.FEATURE_DATA, target_path, path)
+        if not os.path.exists(feature_path):
+            os.makedirs(feature_path)
+        for obj in obj_list.obj_list:
+            total_num = np.prod(obj.num)
+            for i in range(total_num):
+                target_file_name = get_target_file_name(obj.sound_prefix + obj.obj_file_name, obj.start_num + i)
+                raw_file_path = os.path.join(raw_path, target_file_name)
+                feature_file_path = os.path.join(feature_path, target_file_name)
+                if not os.path.exists(feature_file_path) or overwrite:
+                    raw_file_data = np.load(raw_file_path)
+                    raw_file_data = transform(raw_file_data)
+                    np.save(feature_file_path, raw_file_data)
 
-    # 读取实验数据
-    exp_data = []
-    exp_label = []
-    exp_path = os.path.join(global_var.DATASET, global_var.RAW_DATA, 'exp_data')
-    for obj in obj_list.obj_list:
-        total_num = np.prod(obj.num)
-        for i in range(total_num):
-            exp_file_name = get_target_file_name(obj.sound_prefix + obj.obj_file_name, obj.start_num + i)
-            exp_file_path = os.path.join(exp_path, exp_file_name)
-            exp_file_data = np.load(exp_file_path)
-            exp_data.append(exp_file_data)
-            label = get_label(exp_file_name)
-            exp_label.append(label)
 
-    dataset = [sim_data, sim_label, exp_data, exp_label]
+# 载入feature_data数据
+def load_feature_data(obj_list, target_path):
+    sim_data, sim_labels = [], []
+    exp_data, exp_labels = [], []
+    for path in ['sim_data', 'target_data']:
+        feature_path = os.path.join(global_var.DATASET, global_var.FEATURE_DATA, target_path, path)
+        for obj in obj_list.obj_list:
+            total_num = np.prod(obj.num)
+            for i in range(total_num):
+                target_file_name = get_target_file_name(obj.sound_prefix + obj.obj_file_name, obj.start_num + i)
+                feature_file_path = os.path.join(feature_path, target_file_name)
+                feature_file_data = np.load(feature_file_path)
+                feature_file_label = get_label(target_file_name)
+                if path == 'sim-data':
+                    sim_data.append(feature_file_data)
+                    sim_labels.append(feature_file_label)
+                else:
+                    exp_data.append(feature_file_data)
+                    exp_labels.append(feature_file_label)
+
+    sim_data, sim_labels = np.array(sim_data), np.array(sim_labels)
+    exp_data, exp_labels = np.array(exp_data), np.array(exp_labels)
+    sim_data, sim_labels = shuffle(sim_data, sim_labels, random_state=42)
+    exp_data, exp_labels = shuffle(exp_data, exp_labels, random_state=42)
+    dataset = [sim_data, sim_labels, exp_data, exp_labels]
     
     return dataset
 
 
-def get_train_test_dataset(dataset):
+def get_label_set(labels):
+    labels_set = set(labels)
+    return labels_set.__len__()
+
+
+# 根据数据类别进行分类，包含所有类别，每类的一部分作为训练集，一部分作为测试集
+def get_train_test_dataset_1(dataset, ratio):
+    sim_data, sim_label, exp_data, exp_label = dataset
+    data_all = np.vstack(sim_data, exp_data)
+    labels_all = np.hstack(sim_label, exp_label)
+    labels_set = set(labels_all)
+
+    train_data = np.empty((0, 10, 10), dtype=np.float32)
+    train_labels = np.empty((0,), np.int32)
+    test_data = np.empty((0, 10, 10), dtype=np.float32)
+    test_labels = np.empty((0,), np.int32)
+
+    for label in labels_set:
+        index = labels_all == label
+        num = int(np.sum(index) * ratio)
+        data = data_all[index]
+        labels = labels_all[index]
+        train_data = np.vstack(train_data, data[:num])
+        train_labels = np.hstack(train_labels, labels[:num])
+        test_data = np.vstack(test_data, data[num:])
+        test_labels = np.hstack(test_data, labels[num:])
+
+    train_dataset = KnockDataset(train_data, train_labels)
+    test_dataset = KnockDataset(test_data, test_labels)
+
+    return train_dataset, test_dataset
+
+
+# 将部分标签划分出来
+def get_train_test_dataset_2(dataset, n):
     train_dataset, test_dataset = 0, 0
     return train_dataset, test_dataset
 
 
+def get_train_test_dataset_3(dataset, n):
+    pass
+
+
+class KnockDataset(Dataset):
+
+    def __init__(self, data, labels, transform=None, target_transform=None):
+        self.knock_data = data
+        self.knock_labels = labels
+        self.transform = transform
+        self.target_transform = target_transform
+
+        self.label_set = set(self.knock_labels)
+        self.n_classes = self.val_label_set.__len__()
+
+        self.knock_data = torch.Tensor(self.knock_data)
+        self.knock_labels = torch.Tensor(self.knock_labels)
+
+        if torch.cuda.is_available():
+            self.knock_data = self.knock_data.cuda()
+            self.knock_labels = self.knock_labels.cuda()
+
+    def __getitem__(self, index):
+        knock_data = self.knock_data[index]
+        knock_label = int(self.knock_labels[index])
+        if self.transform:
+            knock_data = self.transform(knock_data)
+        if self.target_transform:
+            knock_label = self.target_transform(knock_label)
+        return knock_data, knock_label
+
+    def __len__(self):
+        return len(self.knock_labels)
+
+
 if __name__ == '__main__':
     name = 'exp'
+    target_path = 'fbank_dnoised_data'
+    feature_transform = fbank_transform
     obj_list = ObjList(name)
-    copy_file(obj_list)
-    dataset = get_dataset(obj_list)
-    train_dataset, test_dataset = get_train_test_dataset(dataset)
+    copy_sim_file(obj_list)
+    raw_to_feature_dataset(obj_list, target_path, feature_transform)
+    dataset = load_feature_data(obj_list, target_path)
+    train_dataset, test_dataset = get_train_test_dataset_1(dataset, 0.8)
